Create Config.ipython
The file is intended to be used as two separate small modules in Python programs to check the system properties and load relevant hardware-based configurations. It is mainly designed to check for Intel sklearn patch and apply it if CPU configuration allows. In the case of parallel CPU processing, it also can check if the CPU architecture will benefit from parallel versions of other modules.